+++ tests/functional/lt_test_rev_ecc_key_store.c
@@ -65,7 +65,7 @@ static lt_ret_t lt_test_rev_ecc_key_store_cleanup(void)
         }
 
         LT_LOG_INFO("Reading slot #%" PRIu8 " (should fail)", i);
-        ret = lt_ecc_key_read(g_h, i, read_pub_key, &curve, &origin);
+        ret = lt_ecc_key_read(g_h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin);
         if (LT_L3_ECC_INVALID_KEY != ret) {
             LT_LOG_ERROR("Return value is not LT_L3_ECC_INVALID_KEY.");
             return ret;
@@ -120,7 +120,8 @@ void lt_test_rev_ecc_key_store(lt_handle_t *h)
         LT_LOG_INFO("Testing ECC key slot #%" PRIu8 "...", i);
 
         LT_LOG_INFO("Checking if slot is empty...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Storing invalid private key (should fail)...");
         LT_TEST_ASSERT(LT_L3_FAIL, lt_ecc_key_store(h, i, TR01_CURVE_P256, p256_invalid_priv_test_key));
@@ -135,7 +136,7 @@ void lt_test_rev_ecc_key_store(lt_handle_t *h)
         LT_TEST_ASSERT(LT_L3_FAIL, lt_ecc_key_store(h, i, TR01_CURVE_ED25519, ed25519_priv_test_key));
 
         LT_LOG_INFO("Reading the stored public key...");
-        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Checking curve type of the read key...");
         LT_TEST_ASSERT(1, (curve == TR01_CURVE_P256));
@@ -157,7 +158,8 @@ void lt_test_rev_ecc_key_store(lt_handle_t *h)
         LT_LOG_INFO("Testing ECC key slot #%" PRIu8 "...", i);
 
         LT_LOG_INFO("Checking if slot is empty...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Storing private key pre-generated using Ed25519 curve...");
         LT_TEST_ASSERT(LT_OK, lt_ecc_key_store(h, i, TR01_CURVE_ED25519, ed25519_priv_test_key));
@@ -169,7 +171,7 @@ void lt_test_rev_ecc_key_store(lt_handle_t *h)
         LT_TEST_ASSERT(LT_L3_FAIL, lt_ecc_key_store(h, i, TR01_CURVE_P256, p256_priv_test_key));
 
         LT_LOG_INFO("Reading the stored public key...");
-        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_OK, lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
 
         LT_LOG_INFO("Checking curve type of the read key...");
         LT_TEST_ASSERT(1, (curve == TR01_CURVE_ED25519));
@@ -184,7 +186,8 @@ void lt_test_rev_ecc_key_store(lt_handle_t *h)
         LT_TEST_ASSERT(LT_OK, lt_ecc_key_erase(h, i));
 
         LT_LOG_INFO("Trying to read the erased slot (should fail)...");
-        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY, lt_ecc_key_read(h, i, read_pub_key, &curve, &origin));
+        LT_TEST_ASSERT(LT_L3_ECC_INVALID_KEY,
+                       lt_ecc_key_read(h, i, read_pub_key, sizeof(read_pub_key), &curve, &origin));
     }
     LT_LOG_LINE();