+++ CHANGELOG.md
@@ -108,6 +108,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed other macros and definitions used internally.
 - Changed type of the first parameter of `lt_random_bytes` to `lt_handle_t`.
 - Replaced various numeric constants with appropriate macro constants and sizeof operators.
+- `lt_get_info_fw_bank()`: Renamed parameter `max_len` to `max_size`.
+- `lt_get_log_req()`: Renamed parameter `log_msg_len` to `read_size`.
+- `lt_r_mem_data_read()`: Renamed parameter `size` to `read_size`.
 
 ### Added
 - CMake option for setting logging verbosity level: `LT_LOG_LVL`.
@@ -118,6 +121,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Constants for length of P256 and ED25519 private and public keys.
 - Constant for length of ECDSA/EDDSA signature.
 - Constants for lengths of L2 request fields (ID, LEN, CRC).
+- `lt_get_info_fw_bank()`: Parameter `read_size` to indicate the number of read bytes from the FW bank.
+- `lt_get_log_req()`: Parameter `max_size` to check whether the output buffer is big enough.
+- `lt_r_mem_data_read()`: Parameter `max_size` to check whether the output buffer is big enough.
+- `lt_ecc_key_read()`: Parameter `max_size` to check whether the output buffer is big enough.
 
 ### Fixed
 - `lt_r_mem_data_write()`, `lt_out__r_mem_data_write()`: Mark `data` as `const`.
@@ -126,6 +133,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Unused `pairing_key_slot_t` enum.
+- `lt_get_st_pub()`: Parameter `stpub_len`.
 
 ## [1.0.0]